requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
ebus-python 
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
opencv-python==4.9.0.80
Werkzeug==3.0.1